feat: add metrics for the invalid headers cache
